requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.7.4
aiohttp-socks==0.5.5
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asteroid-filterbanks==0.4.0
async-timeout==3.0.1
attrs==25.3.0
av==15.0.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.0
ctranslate2==4.4.0
cycler==0.12.1
deep-translator==1.11.4
distro==1.9.0
docopt==0.6.2
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
faster-whisper==1.2.0
feedparser==6.0.11
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2024.6.1
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
julius==0.2.7
kiwisolver==1.4.7
langchain==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.14
lazy_loader==0.4
lightning==2.5.3
lightning-utilities==0.15.2
lxml==6.0.0
maigret==0.4.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.2.1
ninja==1.13.0
nltk==3.9.1
numpy==2.0.2
omegaconf==2.3.0
onnxruntime==1.19.2
openai==1.99.9
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
optuna==4.4.0
orjson==3.11.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
praw==7.8.1
prawcore==2.4.0
primePy==1.3
propcache==0.3.2
protobuf==6.32.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-socks==2.5.3
pytorch-lightning==2.5.3
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-futures==1.0.2
requests-toolbelt==1.0.0
rich==14.1.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.35.0
semver==3.0.4
sentencepiece==0.2.1
sgmllib3k==1.0.0
shapely==2.0.7
shellingham==1.5.4
# Editable Git install with no remote (sherlock-project==0)
-e /Users/apple/Desktop/osint-llm-tool/tools/sherlock-master
six==1.17.0
sniffio==1.3.1
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
speechbrain==1.0.3
SQLAlchemy==2.0.43
starlette==0.47.2
stem==1.8.2
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2024.8.30
tokenizers==0.21.4
tomli==2.2.1
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchmetrics==1.8.1
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.2
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
whisperx @ https://github.com/m-bain/whisperx/archive/refs/heads/main.zip#sha256=12f779c076266663be75f0f7cf30590b2acbfef2f37e5224cd814f587e5a0eda
wsproto==1.2.0
yarl==1.20.1
youtube-search-python==1.6.6
yt-dlp==2025.8.11
zipp==3.23.0
zstandard==0.23.0
